+++ apps/dokploy/pages/dashboard/project/[projectId]/environment/[environmentId]/services/mariadb/[mariadbId].tsx
@@ -1,7 +1,6 @@
 import { validateRequest } from "@dokploy/server/lib/auth";
 import { createServerSideHelpers } from "@trpc/react-query/server";
-import copy from "copy-to-clipboard";
-import { GlobeIcon, HelpCircle, ServerOff } from "lucide-react";
+import { HelpCircle, ServerOff } from "lucide-react";
 import type {
 	GetServerSidePropsContext,
 	InferGetServerSidePropsType,
@@ -9,29 +8,20 @@ import type {
 import Head from "next/head";
 import Link from "next/link";
 import { useRouter } from "next/router";
-import { type ReactElement, useEffect, useState } from "react";
-import { toast } from "sonner";
+import { type ReactElement, useState } from "react";
 import superjson from "superjson";
-import { ShowClusterSettings } from "@/components/dashboard/application/advanced/cluster/show-cluster-settings";
-import { AddCommand } from "@/components/dashboard/application/advanced/general/add-command";
-import { ShowPorts } from "@/components/dashboard/application/advanced/ports/show-port";
-import { ShowRedirects } from "@/components/dashboard/application/advanced/redirects/show-redirects";
-import { ShowSecurity } from "@/components/dashboard/application/advanced/security/show-security";
-import { ShowResources } from "@/components/dashboard/application/advanced/show-resources";
-import { ShowTraefikConfig } from "@/components/dashboard/application/advanced/traefik/show-traefik-config";
-import { ShowVolumes } from "@/components/dashboard/application/advanced/volumes/show-volumes";
-import { ShowDeployments } from "@/components/dashboard/application/deployments/show-deployments";
-import { ShowDomains } from "@/components/dashboard/application/domains/show-domains";
-import { ShowEnvironment } from "@/components/dashboard/application/environment/show";
-import { ShowGeneralApplication } from "@/components/dashboard/application/general/show";
+import { ShowEnvironment } from "@/components/dashboard/application/environment/show-enviroment";
 import { ShowDockerLogs } from "@/components/dashboard/application/logs/show";
-import { ShowPreviewDeployments } from "@/components/dashboard/application/preview-deployments/show-preview-deployments";
-import { ShowSchedules } from "@/components/dashboard/application/schedules/show-schedules";
-import { UpdateApplication } from "@/components/dashboard/application/update-application";
-import { ShowVolumeBackups } from "@/components/dashboard/application/volume-backups/show-volume-backups";
 import { DeleteService } from "@/components/dashboard/compose/delete-service";
+import { ShowBackups } from "@/components/dashboard/database/backups/show-backups";
+import { ShowExternalMariadbCredentials } from "@/components/dashboard/mariadb/general/show-external-mariadb-credentials";
+import { ShowGeneralMariadb } from "@/components/dashboard/mariadb/general/show-general-mariadb";
+import { ShowInternalMariadbCredentials } from "@/components/dashboard/mariadb/general/show-internal-mariadb-credentials";
+import { UpdateMariadb } from "@/components/dashboard/mariadb/update-mariadb";
 import { ContainerFreeMonitoring } from "@/components/dashboard/monitoring/free/container/show-free-container-monitoring";
 import { ContainerPaidMonitoring } from "@/components/dashboard/monitoring/paid/container/show-paid-container-monitoring";
+import { ShowDatabaseAdvancedSettings } from "@/components/dashboard/shared/show-database-advanced-settings";
+import { MariadbIcon } from "@/components/icons/data-tools-icons";
 import { DashboardLayout } from "@/components/layouts/dashboard-layout";
 import { BreadcrumbSidebar } from "@/components/shared/breadcrumb-sidebar";
 import { StatusTooltip } from "@/components/shared/status-tooltip";
@@ -52,47 +42,29 @@ import {
 	TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { UseKeyboardNav } from "@/hooks/use-keyboard-nav";
+import { cn } from "@/lib/utils";
 import { appRouter } from "@/server/api/root";
 import { api } from "@/utils/api";
 
-type TabState =
-	| "projects"
-	| "settings"
-	| "advanced"
-	| "deployments"
-	| "domains"
-	| "monitoring"
-	| "preview-deployments"
-	| "volume-backups";
+type TabState = "projects" | "monitoring" | "settings" | "backups" | "advanced";
 
-const Service = (
+const Mariadb = (
 	props: InferGetServerSidePropsType<typeof getServerSideProps>,
 ) => {
 	const [_toggleMonitoring, _setToggleMonitoring] = useState(false);
-	const { applicationId, activeTab } = props;
+
+	const { mariadbId, activeTab } = props;
 	const router = useRouter();
-	const { projectId } = router.query;
-	const [tab, setTab] = useState<TabState>(activeTab);
-
-	useEffect(() => {
-		if (router.query.tab) {
-			setTab(router.query.tab as TabState);
-		}
-	}, [router.query.tab]);
-
-	const { data } = api.application.one.useQuery(
-		{ applicationId },
-		{
-			refetchInterval: 5000,
-		},
-	);
+	const { projectId, environmentId } = router.query;
+	const [tab, setSab] = useState<TabState>(activeTab);
+	const { data } = api.mariadb.one.useQuery({ mariadbId });
+	const { data: auth } = api.user.get.useQuery();
 
 	const { data: isCloud } = api.settings.isCloud.useQuery();
-	const { data: auth } = api.user.get.useQuery();
 
 	return (
 		<div className="pb-10">
-			<UseKeyboardNav forPage="application" />
+			<UseKeyboardNav forPage="mariadb" />
 			<BreadcrumbSidebar
 				list={[
 					{ name: "Projects", href: "/dashboard/projects" },
@@ -102,27 +74,27 @@ const Service = (
 					},
 					{
 						name: data?.name || "",
-						href: `/dashboard/project/${projectId}/services/application/${applicationId}`,
+						href: `/dashboard/project/${projectId}/environment/${environmentId}/services/mariadb/${mariadbId}`,
 					},
 				]}
 			/>
-			<Head>
-				<title>
-					Application: {data?.name} - {data?.project.name} | Dokploy
-				</title>
-			</Head>
-			<div className="w-full">
-				<Card className="h-full bg-sidebar p-2.5 rounded-xl w-full">
+			<div className="flex flex-col gap-4">
+				<Head>
+					<title>
+						Database: {data?.name} - {data?.project.name} | Dokploy
+					</title>
+				</Head>
+				<Card className="h-full bg-sidebar  p-2.5 rounded-xl w-full">
 					<div className="rounded-xl bg-background shadow-md ">
 						<CardHeader className="flex flex-row justify-between items-center">
 							<div className="flex flex-col">
 								<CardTitle className="text-xl flex flex-row gap-2">
 									<div className="relative flex flex-row gap-4">
-										<div className="absolute -right-1 -top-2">
+										<div className="absolute -right-1  -top-2">
 											<StatusTooltip status={data?.applicationStatus} />
 										</div>
 
-										<GlobeIcon className="h-6 w-6 text-muted-foreground" />
+										<MariadbIcon className="h-6 w-6 text-muted-foreground" />
 									</div>
 									{data?.name}
 								</CardTitle>
@@ -137,13 +109,6 @@ const Service = (
 							<div className="flex flex-col h-fit w-fit gap-2">
 								<div className="flex flex-row h-fit w-fit gap-2">
 									<Badge
-										className="cursor-pointer"
-										onClick={() => {
-											if (data?.server?.ipAddress) {
-												copy(data.server.ipAddress);
-												toast.success("IP Address Copied!");
-											}
-										}}
 										variant={
 											!data?.serverId
 												? "default"
@@ -177,11 +142,10 @@ const Service = (
 										</TooltipProvider>
 									)}
 								</div>
-
 								<div className="flex flex-row gap-2 justify-end">
-									<UpdateApplication applicationId={applicationId} />
+									<UpdateMariadb mariadbId={mariadbId} />
 									{(auth?.role === "owner" || auth?.canDeleteServices) && (
-										<DeleteService id={applicationId} type="application" />
+										<DeleteService id={mariadbId} type="mariadb" />
 									)}
 								</div>
 							</div>
@@ -214,43 +178,46 @@ const Service = (
 									defaultValue="general"
 									className="w-full"
 									onValueChange={(e) => {
-										setTab(e as TabState);
-										const newPath = `/dashboard/project/${projectId}/services/application/${applicationId}?tab=${e}`;
-										router.push(newPath);
+										setSab(e as TabState);
+										const newPath = `/dashboard/project/${projectId}/environment/${environmentId}/services/mariadb/${mariadbId}?tab=${e}`;
+
+										router.push(newPath, undefined, { shallow: true });
 									}}
 								>
-									<div className="flex flex-row items-center justify-between w-full overflow-auto">
-										<TabsList className="flex gap-8 max-md:gap-4 justify-start">
+									<div className="flex flex-row items-center justify-between w-full gap-4 overflow-x-scroll">
+										<TabsList
+											className={cn(
+												"md:grid md:w-fit max-md:overflow-y-scroll justify-start",
+												isCloud && data?.serverId
+													? "md:grid-cols-6"
+													: data?.serverId
+														? "md:grid-cols-5"
+														: "md:grid-cols-6",
+											)}
+										>
 											<TabsTrigger value="general">General</TabsTrigger>
 											<TabsTrigger value="environment">Environment</TabsTrigger>
-											<TabsTrigger value="domains">Domains</TabsTrigger>
-											<TabsTrigger value="preview-deployments">
-												Preview Deployments
-											</TabsTrigger>
-											<TabsTrigger value="schedules">Schedules</TabsTrigger>
-											<TabsTrigger value="volume-backups">
-												Volume Backups
-											</TabsTrigger>
-											<TabsTrigger value="deployments">Deployments</TabsTrigger>
 											<TabsTrigger value="logs">Logs</TabsTrigger>
 											{((data?.serverId && isCloud) || !data?.server) && (
 												<TabsTrigger value="monitoring">Monitoring</TabsTrigger>
 											)}
+											<TabsTrigger value="backups">Backups</TabsTrigger>
 											<TabsTrigger value="advanced">Advanced</TabsTrigger>
 										</TabsList>
 									</div>
 
 									<TabsContent value="general">
 										<div className="flex flex-col gap-4 pt-2.5">
-											<ShowGeneralApplication applicationId={applicationId} />
+											<ShowGeneralMariadb mariadbId={mariadbId} />
+											<ShowInternalMariadbCredentials mariadbId={mariadbId} />
+											<ShowExternalMariadbCredentials mariadbId={mariadbId} />
 										</div>
 									</TabsContent>
 									<TabsContent value="environment">
 										<div className="flex flex-col gap-4 pt-2.5">
-											<ShowEnvironment applicationId={applicationId} />
+											<ShowEnvironment id={mariadbId} type="mariadb" />
 										</div>
 									</TabsContent>
-
 									<TabsContent value="monitoring">
 										<div className="pt-2.5">
 											<div className="flex flex-col gap-4 border rounded-lg p-6">
@@ -264,21 +231,19 @@ const Service = (
 													/>
 												) : (
 													<>
-														{/* {monitoring?.enabledFeatures &&
-															isCloud &&
-															data?.serverId && (
-																<div className="flex flex-row border w-fit p-4 rounded-lg items-center gap-2">
-																	<Label className="text-muted-foreground">
-																		Change Monitoring
-																	</Label>
-																	<Switch
-																		checked={toggleMonitoring}
-																		onCheckedChange={setToggleMonitoring}
-																	/>
-																</div>
-															)} */}
+														{/* {monitoring?.enabledFeatures && (
+															<div className="flex flex-row border w-fit p-4 rounded-lg items-center gap-2">
+																<Label className="text-muted-foreground">
+																	Change Monitoring
+																</Label>
+																<Switch
+																	checked={toggleMonitoring}
+																	onCheckedChange={setToggleMonitoring}
+																/>
+															</div>
+														)}
 
-														{/* {toggleMonitoring ? (
+														{toggleMonitoring ? (
 															<ContainerPaidMonitoring
 																appName={data?.appName || ""}
 																baseUrl={`http://${monitoring?.serverIp}:${monitoring?.metricsConfig?.server?.port}`}
@@ -286,78 +251,37 @@ const Service = (
 																	monitoring?.metricsConfig?.server?.token || ""
 																}
 															/>
-														) : ( */}
-														<div>
-															<ContainerFreeMonitoring
-																appName={data?.appName || ""}
-															/>
-														</div>
+														) : (
+															<div> */}
+														<ContainerFreeMonitoring
+															appName={data?.appName || ""}
+														/>
+														{/* </div> */}
 														{/* )} */}
 													</>
 												)}
 											</div>
 										</div>
 									</TabsContent>
-
 									<TabsContent value="logs">
-										<div className="flex flex-col gap-4 pt-2.5">
+										<div className="flex flex-col gap-4  pt-2.5">
 											<ShowDockerLogs
+												serverId={data?.serverId || ""}
 												appName={data?.appName || ""}
-												serverId={data?.serverId || ""}
 											/>
 										</div>
 									</TabsContent>
-									<TabsContent value="schedules">
+									<TabsContent value="backups">
 										<div className="flex flex-col gap-4 pt-2.5">
-											<ShowSchedules
-												id={applicationId}
-												scheduleType="application"
-											/>
-										</div>
-									</TabsContent>
-									<TabsContent value="deployments" className="w-full pt-2.5">
-										<div className="flex flex-col gap-4 border rounded-lg">
-											<ShowDeployments
-												id={applicationId}
-												type="application"
-												serverId={data?.serverId || ""}
-												refreshToken={data?.refreshToken || ""}
-											/>
-										</div>
-									</TabsContent>
-									<TabsContent value="volume-backups" className="w-full pt-2.5">
-										<div className="flex flex-col gap-4 border rounded-lg">
-											<ShowVolumeBackups
-												id={applicationId}
-												type="application"
-												serverId={data?.serverId || ""}
-											/>
-										</div>
-									</TabsContent>
-									<TabsContent value="preview-deployments" className="w-full">
-										<div className="flex flex-col gap-4 pt-2.5">
-											<ShowPreviewDeployments applicationId={applicationId} />
-										</div>
-									</TabsContent>
-									<TabsContent value="domains" className="w-full">
-										<div className="flex flex-col gap-4 pt-2.5">
-											<ShowDomains id={applicationId} type="application" />
+											<ShowBackups id={mariadbId} databaseType="mariadb" />
 										</div>
 									</TabsContent>
 									<TabsContent value="advanced">
 										<div className="flex flex-col gap-4 pt-2.5">
-											<AddCommand applicationId={applicationId} />
-											<ShowClusterSettings
-												id={applicationId}
-												type="application"
+											<ShowDatabaseAdvancedSettings
+												id={mariadbId}
+												type="mariadb"
 											/>
-
-											<ShowResources id={applicationId} type="application" />
-											<ShowVolumes id={applicationId} type="application" />
-											<ShowRedirects applicationId={applicationId} />
-											<ShowSecurity applicationId={applicationId} />
-											<ShowPorts applicationId={applicationId} />
-											<ShowTraefikConfig applicationId={applicationId} />
 										</div>
 									</TabsContent>
 								</Tabs>
@@ -370,20 +294,17 @@ const Service = (
 	);
 };
 
-export default Service;
-Service.getLayout = (page: ReactElement) => {
+export default Mariadb;
+Mariadb.getLayout = (page: ReactElement) => {
 	return <DashboardLayout>{page}</DashboardLayout>;
 };
 
 export async function getServerSideProps(
-	ctx: GetServerSidePropsContext<{
-		applicationId: string;
-		activeTab: TabState;
-	}>,
+	ctx: GetServerSidePropsContext<{ mariadbId: string; activeTab: TabState; environmentId: string }>,
 ) {
 	const { query, params, req, res } = ctx;
-
 	const activeTab = query.tab;
+
 	const { user, session } = await validateRequest(req);
 	if (!user) {
 		return {
@@ -406,20 +327,18 @@ export async function getServerSideProps(
 		transformer: superjson,
 	});
 
-	// Valid project, if not return to initial homepage....
-	if (typeof params?.applicationId === "string") {
+	if (typeof params?.mariadbId === "string") {
 		try {
-			await helpers.application.one.fetch({
-				applicationId: params?.applicationId,
+			await helpers.mariadb.one.fetch({
+				mariadbId: params?.mariadbId,
 			});
-
 			await helpers.settings.isCloud.prefetch();
-
 			return {
 				props: {
 					trpcState: helpers.dehydrate(),
-					applicationId: params?.applicationId,
+					mariadbId: params?.mariadbId,
 					activeTab: (activeTab || "general") as TabState,
+					environmentId: params?.environmentId,
 				},
 			};
 		} catch {
