+++ apps/dokploy/pages/dashboard/project/[projectId]/environment/[environmentId]/services/postgres/[postgresId].tsx
@@ -13,14 +13,15 @@ import superjson from "superjson";
 import { ShowEnvironment } from "@/components/dashboard/application/environment/show-enviroment";
 import { ShowDockerLogs } from "@/components/dashboard/application/logs/show";
 import { DeleteService } from "@/components/dashboard/compose/delete-service";
+import { ShowBackups } from "@/components/dashboard/database/backups/show-backups";
 import { ContainerFreeMonitoring } from "@/components/dashboard/monitoring/free/container/show-free-container-monitoring";
 import { ContainerPaidMonitoring } from "@/components/dashboard/monitoring/paid/container/show-paid-container-monitoring";
-import { ShowExternalRedisCredentials } from "@/components/dashboard/redis/general/show-external-redis-credentials";
-import { ShowGeneralRedis } from "@/components/dashboard/redis/general/show-general-redis";
-import { ShowInternalRedisCredentials } from "@/components/dashboard/redis/general/show-internal-redis-credentials";
-import { UpdateRedis } from "@/components/dashboard/redis/update-redis";
+import { ShowExternalPostgresCredentials } from "@/components/dashboard/postgres/general/show-external-postgres-credentials";
+import { ShowGeneralPostgres } from "@/components/dashboard/postgres/general/show-general-postgres";
+import { ShowInternalPostgresCredentials } from "@/components/dashboard/postgres/general/show-internal-postgres-credentials";
+import { UpdatePostgres } from "@/components/dashboard/postgres/update-postgres";
 import { ShowDatabaseAdvancedSettings } from "@/components/dashboard/shared/show-database-advanced-settings";
-import { RedisIcon } from "@/components/icons/data-tools-icons";
+import { PostgresqlIcon } from "@/components/icons/data-tools-icons";
 import { DashboardLayout } from "@/components/layouts/dashboard-layout";
 import { BreadcrumbSidebar } from "@/components/shared/breadcrumb-sidebar";
 import { StatusTooltip } from "@/components/shared/status-tooltip";
@@ -45,25 +46,24 @@ import { cn } from "@/lib/utils";
 import { appRouter } from "@/server/api/root";
 import { api } from "@/utils/api";
 
-type TabState = "projects" | "monitoring" | "settings" | "advanced";
+type TabState = "projects" | "monitoring" | "settings" | "backups" | "advanced";
 
-const Redis = (
+const Postgresql = (
 	props: InferGetServerSidePropsType<typeof getServerSideProps>,
 ) => {
 	const [_toggleMonitoring, _setToggleMonitoring] = useState(false);
-	const { redisId, activeTab } = props;
+	const { postgresId, activeTab } = props;
 	const router = useRouter();
-	const { projectId } = router.query;
+	const { projectId, environmentId } = router.query;
 	const [tab, setSab] = useState<TabState>(activeTab);
-	const { data } = api.redis.one.useQuery({ redisId });
-
+	const { data } = api.postgres.one.useQuery({ postgresId });
 	const { data: auth } = api.user.get.useQuery();
 
 	const { data: isCloud } = api.settings.isCloud.useQuery();
 
 	return (
 		<div className="pb-10">
-			<UseKeyboardNav forPage="redis" />
+			<UseKeyboardNav forPage="postgres" />
 			<BreadcrumbSidebar
 				list={[
 					{ name: "Projects", href: "/dashboard/projects" },
@@ -73,7 +73,7 @@ const Redis = (
 					},
 					{
 						name: data?.name || "",
-						href: `/dashboard/project/${projectId}/services/redis/${redisId}`,
+						href: `/dashboard/project/${projectId}/environment/${environmentId}/services/postgres/${postgresId}`,
 					},
 				]}
 			/>
@@ -93,7 +93,7 @@ const Redis = (
 											<StatusTooltip status={data?.applicationStatus} />
 										</div>
 
-										<RedisIcon className="h-6 w-6 text-muted-foreground" />
+										<PostgresqlIcon className="h-6 w-6 text-muted-foreground" />
 									</div>
 									{data?.name}
 								</CardTitle>
@@ -143,9 +143,9 @@ const Redis = (
 								</div>
 
 								<div className="flex flex-row gap-2 justify-end">
-									<UpdateRedis redisId={redisId} />
+									<UpdatePostgres postgresId={postgresId} />
 									{(auth?.role === "owner" || auth?.canDeleteServices) && (
-										<DeleteService id={redisId} type="redis" />
+										<DeleteService id={postgresId} type="postgres" />
 									)}
 								</div>
 							</div>
@@ -179,7 +179,7 @@ const Redis = (
 									className="w-full"
 									onValueChange={(e) => {
 										setSab(e as TabState);
-										const newPath = `/dashboard/project/${projectId}/services/redis/${redisId}?tab=${e}`;
+										const newPath = `/dashboard/project/${projectId}/environment/${environmentId}/services/postgres/${postgresId}?tab=${e}`;
 
 										router.push(newPath, undefined, { shallow: true });
 									}}
@@ -189,10 +189,10 @@ const Redis = (
 											className={cn(
 												"md:grid md:w-fit max-md:overflow-y-scroll justify-start",
 												isCloud && data?.serverId
-													? "md:grid-cols-5"
+													? "md:grid-cols-6"
 													: data?.serverId
-														? "md:grid-cols-4"
-														: "md:grid-cols-5",
+														? "md:grid-cols-5"
+														: "md:grid-cols-6",
 											)}
 										>
 											<TabsTrigger value="general">General</TabsTrigger>
@@ -201,20 +201,25 @@ const Redis = (
 											{((data?.serverId && isCloud) || !data?.server) && (
 												<TabsTrigger value="monitoring">Monitoring</TabsTrigger>
 											)}
+											<TabsTrigger value="backups">Backups</TabsTrigger>
 											<TabsTrigger value="advanced">Advanced</TabsTrigger>
 										</TabsList>
 									</div>
 
 									<TabsContent value="general">
 										<div className="flex flex-col gap-4 pt-2.5">
-											<ShowGeneralRedis redisId={redisId} />
-											<ShowInternalRedisCredentials redisId={redisId} />
-											<ShowExternalRedisCredentials redisId={redisId} />
+											<ShowGeneralPostgres postgresId={postgresId} />
+											<ShowInternalPostgresCredentials
+												postgresId={postgresId}
+											/>
+											<ShowExternalPostgresCredentials
+												postgresId={postgresId}
+											/>
 										</div>
 									</TabsContent>
 									<TabsContent value="environment">
 										<div className="flex flex-col gap-4 pt-2.5">
-											<ShowEnvironment id={redisId} type="redis" />
+											<ShowEnvironment id={postgresId} type="postgres" />
 										</div>
 									</TabsContent>
 									<TabsContent value="monitoring">
@@ -230,33 +235,9 @@ const Redis = (
 													/>
 												) : (
 													<>
-														{/* {monitoring?.enabledFeatures && (
-															<div className="flex flex-row border w-fit p-4 rounded-lg items-center gap-2">
-																<Label className="text-muted-foreground">
-																	Change Monitoring
-																</Label>
-																<Switch
-																	checked={toggleMonitoring}
-																	onCheckedChange={setToggleMonitoring}
-																/>
-															</div>
-														)}
-
-														{toggleMonitoring ? (
-															<ContainerPaidMonitoring
-																appName={data?.appName || ""}
-																baseUrl={`http://${monitoring?.serverIp}:${monitoring?.metricsConfig?.server?.port}`}
-																token={
-																	monitoring?.metricsConfig?.server?.token || ""
-																}
-															/>
-														) : (
-															<div> */}
 														<ContainerFreeMonitoring
 															appName={data?.appName || ""}
 														/>
-														{/* </div> */}
-														{/* )} */}
 													</>
 												)}
 											</div>
@@ -270,9 +251,21 @@ const Redis = (
 											/>
 										</div>
 									</TabsContent>
+									<TabsContent value="backups">
+										<div className="flex flex-col gap-4 pt-2.5">
+											<ShowBackups
+												id={postgresId}
+												databaseType="postgres"
+												backupType="database"
+											/>
+										</div>
+									</TabsContent>
 									<TabsContent value="advanced">
 										<div className="flex flex-col gap-4 pt-2.5">
-											<ShowDatabaseAdvancedSettings id={redisId} type="redis" />
+											<ShowDatabaseAdvancedSettings
+												id={postgresId}
+												type="postgres"
+											/>
 										</div>
 									</TabsContent>
 								</Tabs>
@@ -285,17 +278,16 @@ const Redis = (
 	);
 };
 
-export default Redis;
-Redis.getLayout = (page: ReactElement) => {
+export default Postgresql;
+Postgresql.getLayout = (page: ReactElement) => {
 	return <DashboardLayout>{page}</DashboardLayout>;
 };
 
 export async function getServerSideProps(
-	ctx: GetServerSidePropsContext<{ redisId: string; activeTab: TabState }>,
+	ctx: GetServerSidePropsContext<{ postgresId: string; activeTab: TabState; environmentId: string }>,
 ) {
 	const { query, params, req, res } = ctx;
 	const activeTab = query.tab;
-
 	const { user, session } = await validateRequest(req);
 	if (!user) {
 		return {
@@ -317,16 +309,17 @@ export async function getServerSideProps(
 		},
 		transformer: superjson,
 	});
-	if (typeof params?.redisId === "string") {
+
+	if (typeof params?.postgresId === "string") {
 		try {
-			await helpers.redis.one.fetch({
-				redisId: params?.redisId,
+			await helpers.postgres.one.fetch({
+				postgresId: params?.postgresId,
 			});
 			await helpers.settings.isCloud.prefetch();
 			return {
 				props: {
 					trpcState: helpers.dehydrate(),
-					redisId: params?.redisId,
+					postgresId: params?.postgresId,
 					activeTab: (activeTab || "general") as TabState,
 				},
 			};
